+++ spec/hcloud/resources/network_spec.rb
@@ -15,6 +15,8 @@ RSpec.describe HCloud::Network, integration: true, order: :defined do
 
     expect(network.ip_range).to eq "10.0.0.0/16"
 
+    expect(network).not to be_expose_routes_to_vswitch
+
     expect(network.subnets.first.type).to eq "cloud"
     expect(network.subnets.first.ip_range).to eq "10.0.1.0/24"
     expect(network.subnets.first.network_zone).to eq "eu-central"